fix(terrain): cap the DEM texture cache and widen memTextures to int64
The persisted abort message from the emulator finally explained the
Texture2DPool 'memTextures >= 0' crashes: all pool-internal counters
agreed at 2,241,014,164 bytes (2204 live 514x514 DEM textures, 0
reuses, 73 freed) - the pool bookkeeping was always consistent, but
RenderingStats::memTextures is a 32-bit int and wrapped negative past
2GB, tripping the assert. The real defect is RenderTerrain::demTextures
growing without bound: ancestor/descendant retention keeps whole zoom
chains alive while browsing, eventually OOMing the process (the earlier
silent app deaths).

- Track lastUsed on each cached DEM texture and evict least-recently-
  used entries not referenced in the current frame beyond a 96-texture
  budget (~100MB)
- Widen RenderingStats::memTextures to int64_t so the stat cannot wrap
  (and the pool assert is a genuine corruption detector again)

+++ include/mbgl/gfx/rendering_stats.hpp
@@ -73,8 +73,8 @@ struct RenderingStats {
     /// Sum of uniform buffers update sizes
     std::size_t uniformUpdateBytes = 0;
 
-    /// Total texture memory
-    int memTextures = 0;
+    /// Total texture memory (int64_t: terrain scenes can track more than 2GB)
+    int64_t memTextures = 0;
     /// Total buffer memory
     int memBuffers = 0;
     /// Total index buffer memory
